// src/app/interfaces/warehouse-details.ts
export class WarehouseDetail{

  idSiteStore: number;
  idProduct: number;
  warehouseSiteProductCode: string;
  warehouseSiteProductName: string;
  warehouseTotalProducts: number;
  warehouseStockProducts: number;
  warehousePublishedProducts: number;
  warehouseMovingType: string;
  idMovingWarehouseEntryPublished: number;
}